PL-95 Extract interfaces for AuditRecordGenerator
